Merge added io_utils,ecommerce and metrics in notebook

+++ src/metrics.py
@@ -8,7 +8,7 @@ def revenue_per_category(df):
     return (df.groupby("category", dropna=False, observed = True)
                .agg( 
                    total_intäkt= ("revenue", "sum"),                              
-               ).reset_index().astype(int)
+               ).reset_index()
                )
 
 
@@ -46,7 +46,7 @@ def revenue_per_city(df):
     return (df.groupby("city", dropna=False, observed = True)
                .agg( 
                    total_intäkt= ("revenue", "sum"),                                
-               ).reset_index().astype(int)
+               ).reset_index()
                )
 
 
@@ -57,5 +57,5 @@ def change_over_time(df):
     return (df.groupby("month", dropna=False, observed=True)
             .agg(
                 intäkt = ("revenue", "sum"),
-            ).reset_index().astype(int)
+            ).reset_index()
             )
